Adding -tweak:rosapray and fixing flag name
Firstly: most of the changes are to change the slug "rosa_paladin" to
"rosapaladin", because there were a bunch of incorrect flags in places
and this brought things more in line with previous flag slugs.

Anyway, Rosa's Pray command now uses her level as a way to choose
one of nothing/Cure1/Cure2/Cure3/Cure4, randomly. The game rolls RNG
between two level-dependent bounds and compares to various ranges
to determine the spell; the code is in improve_rosa_pray.f4c.
generator.py loads that file. flagspec/etc. has been updated for it.