Fixing bugs with Regen/Bless
There are two bugs with Regen in FE; one is a vanilla FF4 bug, and
the other is to do with axtor-vs-actor references. Both fixes,
which are fairly minimal, are included in the -miscbugfixes flag.
uispec_flagdescriptions was updated, to include this info. See
main_info.md for detailed info about the bugfixes.