Adding -tweak:fusoyaregen
This flag changes Fu's Regen command to restore MP instead of HP
(or just a different amount of MP under the Tellah Maneuver wacky).
The changes are in improve_fusoya_regen_mp.f4c and generator.py;
there's an alert text change in japanese_names.f4c. flagspec/etc
has been updated.

The main challenge with this flag was getting Regen to actually
restore MP; it normally relies on automated application of damage,
so instead we need to manually restore the MP and then get the game
to visually show the healing numbers but not actually do anything
(as well as update the MP dialog).