Modifying Time Is Money and Enogp
The way Enogp used to work is by pretending that all enemies just
gave 0 GP. For the purposes of... other flags, the way it now works
is by running all of the calculations as normal until the game tries
to add the GP spoils to your party GP, at which point it just adds 0
instead. There's also an alert bypass, so the not-actually-gp-reward
text does not show.

Amusingly, Time is Money and Enogp attempted to make the same patch
in the same place. The wacky has now been changed to simply load the
same file that Enogp does, which allows this combination to actually
compile (which matters mostly for random wackies).