Adding -wacky:moneygains
Big Money, Little Gains is a wacky flag that swaps GP and EXP rewards
from battle, as if the reward values for all monsters were swapped.
(So, the various scalings/etc. are handled correctly afterwards.)
Thanks to C&C for this idea.

The way this code change works is by carefully asking the game to
put add monster EXP rewards to the GP spoils field and vice versa;
we need to be careful because FE allows for three-digit boss exp
(hence why randomizer_boss.f4c has been changed), and the -exp:geometric
flag impacts the values that get added to the spoils fields, so we
need to take that into account as well.

This flag is why encounter_no_gp.f4c was changed in a previous commit.