Changing the Kaipo npc to restore map text
In the previous commit, we moved the npc into the abyss to prevent
the player from interacting with "Tory" in the Kaipo Cafe. ... But,
since there's a textbox that we'd like to still see, we can instead
just change the npc to another one that also reads map message 0.
KaipoMan does the trick! So that's what we've done, without moving
the npc at all.