Adding tracker details for BedCid/gated obj items
We added rewards slots for gated objective completion (for Ogated)
and for Cid in the hospital in Dwarf Castle, but we hadn't added the
text that the game can read to show you where you got those items.
That's been changed; simple two-line addition to rewards.f4c
describing the item locations.